move cluster cache code to a separate file
